requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
djangorestframework==3.15.2
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
kombu==5.4.2
mysqlclient==2.2.6
numpy==2.2.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==11.1.0
prompt_toolkit==3.0.48
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
redis==5.2.1
setuptools==75.7.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
